Added eslintrc.json configuration file for Typescript. Converted server.js and utils/config.js to ts.
