+++ server/.eslintrc.json
@@ -1,32 +1,19 @@
 {
-    "root": true,
-    "ignorePatterns": ["**/test", "**/__tests__"],
-    "env": {
-      "node": true,
-      "browser": true,
-      "es2021": true
-    },
-    "plugins": ["react"],
-    "extends": ["eslint:recommended", "plugin:react/recommended"],
-    "parserOptions": {
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "jsx": true
-      }
-    },
-    "rules": {
-        "no-console": "off",
-        "arrow-body-style": "off",
-        "object-curly-newline": "off",
-        "indent": ["warn", 2],
-        "no-unused-vars": ["off", { "vars": "local" }],
-        "prefer-const": "warn",
-        "quotes": ["warn", "single"],
-        "react/prop-types": "off",
-        "semi": ["warn", "always"],
-        "space-infix-ops": "warn"
-    },
-    "settings": {
-      "react": { "version": "detect"}
-    }
+  "env": {
+    "browser": true,
+    "es2021": true
+  },
+  "extends": "standard-with-typescript",
+  "overrides": [],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module",
+    // <PROJECT_NAME>/tsconfig.json style to allow eslint parser to always find the tsconfig no matter where the working file is.
+    "project": "server/tsconfig.json"
+  },
+  "rules": {
+    "no-console": "off",
+    "@typescript-eslint/semi": "off",
+    "@typescript-eslint/comma-dangle": "off"
   }
+}
