+++ server/src/db/investWithFriendsDb.js
