+++ server/src/middleware/auth.js
