+++ server/src/middleware/validator.js
@@ -1,11 +1,18 @@
 const validator = {};
 
+/*
+  Use typescript for this type valication.
+*/
 validator.addHolding = (req, res, next) => {
-	const { user_id, ticker, stock_quantity } = req.body;
-	if (typeof user_id === 'number' && typeof ticker === 'string' && typeof stock_quantity === 'number') {
-		return next();
-	} else {
-		return next(`
+  const { user_id, ticker, stock_quantity } = req.body;
+  if (
+    typeof user_id === 'number' &&
+    typeof ticker === 'string' &&
+    typeof stock_quantity === 'number'
+  ) {
+    return next();
+  } else {
+    return next(`
       error: req.body
       expected: 
         {
@@ -21,15 +28,15 @@ validator.addHolding = (req, res, next) => {
         }
       req.body: ${JSON.stringify(req.body)}
     `);
-	}
+  }
 };
 
 validator.deleteHolding = (req, res, next) => {
-	const { user_id, ticker } = req.body;
-	if (typeof user_id === 'number' && typeof ticker === 'string') {
-		return next();
-	} else {
-		return next(`
+  const { user_id, ticker } = req.body;
+  if (typeof user_id === 'number' && typeof ticker === 'string') {
+    return next();
+  } else {
+    return next(`
       error: req.body
       expected: 
         {
@@ -43,15 +50,15 @@ validator.deleteHolding = (req, res, next) => {
         }
       req.body: ${JSON.stringify(req.body)}
     `);
-	}
+  }
 };
 
 validator.relationships = (req, res, next) => {
   const { user_id, first_name, last_name } = req.body;
-	if (typeof user_id === 'number' && typeof ticker === 'string') {
-		return next();
-	} else {
-		return next(`
+  if (typeof user_id === 'number' && typeof ticker === 'string') {
+    return next();
+  } else {
+    return next(`
       error: req.body
       expected: 
         {
@@ -67,7 +74,7 @@ validator.relationships = (req, res, next) => {
         }
       req.body: ${JSON.stringify(req.body)}
     `);
-	}
-}
+  }
+};
 
 module.exports = validator;
