+++ server/src/server.ts
@@ -1,0 +1,29 @@
+import express from 'express';
+import type { Request, Response, NextFunction, Express } from 'express';
+import cors from 'cors';
+import router from './routes/index';
+
+const app: Express = express();
+const PORT = 4000;
+
+// let cors = require('cors');
+app.use(cors({ credentials: true, origin: 'http://localhost:3000' }));
+
+// import controller objects containing middleware functions
+
+app.use(express.json());
+
+app.use('/api', router);
+
+// global error handling
+/*
+TO-DO: 1. confirm err type. 2. Re-write the err handler to provide more descriptive err message.
+*/
+app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
+  console.log(err);
+  return res.sendStatus(400);
+});
+
+app.listen(PORT, () => {
+  console.log(`Server listening on port: ${PORT}...`);
+});
